added exception handler for not found case, IT test for order API

+++ service-module/src/main/java/com/gmail/evgenymoshchin/service/impl/ItemServiceImpl.java
@@ -1,9 +1,12 @@
 package com.gmail.evgenymoshchin.service.impl;
 
 import com.gmail.evgenymoshchin.repository.ItemRepository;
+import com.gmail.evgenymoshchin.repository.OrderItemRepository;
 import com.gmail.evgenymoshchin.repository.model.Item;
+import com.gmail.evgenymoshchin.repository.model.OrderItem;
 import com.gmail.evgenymoshchin.service.ItemService;
 import com.gmail.evgenymoshchin.service.converters.ItemServiceConverter;
+import com.gmail.evgenymoshchin.service.exception.ServiceException;
 import com.gmail.evgenymoshchin.service.model.ItemDTO;
 import com.gmail.evgenymoshchin.service.model.ItemPageDTO;
 import lombok.RequiredArgsConstructor;
@@ -13,7 +16,10 @@ import javax.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
+import java.util.Objects;
 
+import static com.gmail.evgenymoshchin.service.constants.ExceptionConstants.COPY_OF;
+import static com.gmail.evgenymoshchin.service.constants.ExceptionConstants.ITEM_WAS_NOT_FOUND_MESSAGE;
 import static com.gmail.evgenymoshchin.service.util.ServiceUtil.getNumbersOfPages;
 
 @Service
@@ -21,6 +27,7 @@ import static com.gmail.evgenymoshchin.service.util.ServiceUtil.getNumbersOfPage
 public class ItemServiceImpl implements ItemService {
 
     private final ItemRepository itemRepository;
+    private final OrderItemRepository orderItemRepository;
     private final ItemServiceConverter converter;
 
     @Override
@@ -56,7 +63,11 @@ public class ItemServiceImpl implements ItemService {
     @Transactional
     public ItemDTO findItemById(Long id) {
         Item item = itemRepository.findById(id);
-        return converter.convertItemToDTO(item);
+        if (Objects.nonNull(item)) {
+            return converter.convertItemToDTO(item);
+        } else {
+            throw new ServiceException(String.format(ITEM_WAS_NOT_FOUND_MESSAGE, id));
+        }
     }
 
     @Override
@@ -69,17 +80,31 @@ public class ItemServiceImpl implements ItemService {
     @Transactional
     public void removeItemById(Long id) {
         Item item = itemRepository.findById(id);
-        itemRepository.remove(item);
+        if (Objects.nonNull(item)) {
+            List<OrderItem> orderItems = orderItemRepository.findAll();
+            for (OrderItem orderItem : orderItems) {
+                if (orderItem.getItemId().equals(id)) {
+                    orderItemRepository.remove(orderItem);
+                }
+            }
+            itemRepository.remove(item);
+        } else {
+            throw new ServiceException(String.format(ITEM_WAS_NOT_FOUND_MESSAGE, id));
+        }
     }
 
     @Override
     @Transactional
     public void copyItemById(Long id) {
         Item item = itemRepository.findById(id);
-        Item copiedItem = new Item();
-        copiedItem.setName(item.getName());
-        copiedItem.setPrice(item.getPrice());
-        copiedItem.setDescription(item.getDescription());
-        itemRepository.persist(copiedItem);
+        if (Objects.nonNull(item)) {
+            Item copiedItem = new Item();
+            copiedItem.setName(COPY_OF + item.getName());
+            copiedItem.setPrice(item.getPrice());
+            copiedItem.setDescription(item.getDescription());
+            itemRepository.persist(copiedItem);
+        } else {
+            throw new ServiceException(String.format(ITEM_WAS_NOT_FOUND_MESSAGE, id));
+        }
     }
 }
